requirements.txt
absl-py
easydict
cython
numpy
tensorflow-gpu==2.7.0
tqdm
scipy
pillow
